fix: change location of secret storage
