+++ src/app/content/services/phrase.interceptor.ts
@@ -4,16 +4,16 @@ import {
   HttpRequest,
 } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { BackendService } from './backend.service';
+import { AuthStateService } from '../../core/auth/auth-state.service';
 
 @Injectable()
 export class PhraseInterceptor implements HttpInterceptor {
-  constructor(private api: BackendService) {}
+  constructor(private authStateService: AuthStateService) {}
 
   intercept(request: HttpRequest<any>, next: HttpHandler) {
     request = request.clone({
       setHeaders: {
-        'x-opp-phrase': this.api.secret || '',
+        'x-opp-phrase': this.authStateService.secret || '',
       },
     });
